backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database and ORM
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Supabase client
supabase==2.3.0

# Authentication and JWT
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Pydantic for data validation
pydantic==2.5.0
email-validator==2.2.0

# Environment variables
python-dotenv==1.0.0

# Google APIs and OAuth
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.110.0

# HTTP requests and utilities
requests==2.31.0
httpx==0.25.2

# OpenAI for LLM categorization
openai==1.51.2

# YAML parsing for prompt configuration
PyYAML==6.0.2

# Date and time utilities (for email parsing)
python-dateutil==2.8.2

# Logging and utilities
python-multipart==0.0.6

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-cov==4.1.0
httpx==0.25.2
faker==21.0.0
factory-boy==3.3.0 